// proxy/web/src/app/models/student.model.ts
export class Student {
  id: string;
  firstName: string;
  lastName: string;
  description: string;
  role: string;
  campus: string;
  groupId: string;
  program: string;
  picture: string;
  email: string;
  linkedinLink: string;
  portfolioLink: string;
  skills: Array<string>;
  campusLiterals: string;
  constructor() {}
}